Sonar Issues

+++ src/main/java/ch/css/produkt/kata/bankocr/ZahlenParser.java
@@ -62,7 +62,6 @@ public class ZahlenParser {
     );
 
     public int parseZeichen(NummerZeichen input) {
-        return zeichenMap.get(input).intValue();
-
+        return zeichenMap.get(input);
     }
 }
